requirements.txt
anyio==3.3.3
asgiref==3.4.1
click==8.0.3
colorama==0.4.4
fastapi==0.70.0
h11==0.12.0
idna==3.3
Jinja2==3.0.2
MarkupSafe==2.0.1
Pillow==8.4.0
pydantic==1.8.2
pytesseract==0.3.8
python-multipart==0.0.5
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
tesseract==0.1.3
typing-extensions==3.10.0.2
uvicorn==0.15.0
